backend/requirements.txt
asgiref==3.4.1
astroid==2.9.3
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-cors-headers==3.11.0
django-extensions==3.1.5
django-silk==4.2.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.1
factory-boy==3.2.1
Faker==12.2.0
gprof2dot==2021.2.21
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.4.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.1
pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
